+++ package-source.json
@@ -20,7 +20,7 @@
         "grunt-node-qunit": "",
         "qunit": "",
         "jsdom": "~9.12.0",
-        "jsdom-latest": "npm:jsdom@latest",
+        "jsdom-recent": "npm:jsdom@24.0.0",
         "js-yaml": "~4.1.0",
         "grunt-contrib-sass": "",
         "grunt-contrib-cssmin": "",
