+++ tests/qunit/utils/initPlayer.js
@@ -1,0 +1,10 @@
+const { container } = require('../index.js');
+
+module.exports = (attrs) => {
+    const player = new BetaJS.MediaComponents.VideoPlayer.Dynamics.Player({
+        element: container,
+        attrs: attrs || {}
+    });
+    player.activate();
+    return player;
+}
